requirements.txt
ttkbootstrap
opencv-python
Pillow
pytesseract
numpy
